requirements.txt
sqlalchemy==2.0.32
python-dateutil==2.9.0
requests==2.32.3
